Various cleanup and refactoring.
