+++ irrd/rpki/validators.py
@@ -1,144 +1,21 @@
+import datrie
 from collections import defaultdict
 
 import codecs
-import datrie
-import logging
 import socket
-import textwrap
-import ujson
 from IPy import IP
-from typing import List, Set, Optional, Tuple, Dict
+from typing import Optional, List, Tuple, Set, Dict
 
 from irrd.conf import RPKI_IRR_PSEUDO_SOURCE
-from irrd.rpki.status import RPKIStatus
-from irrd.rpsl.parser import RPSLObject, RPSL_ATTRIBUTE_TEXT_WIDTH
-from irrd.rpsl.rpsl_objects import RPSL_ROUTE_OBJECT_CLASS_FOR_IP_VERSION
 from irrd.storage.database_handler import DatabaseHandler
 from irrd.storage.queries import RPSLDatabaseQuery, ROADatabaseObjectQuery
-from irrd.utils.validators import parse_as_number
+from .importer import ROA
+from .status import RPKIStatus
 
 decode_hex = codecs.getdecoder("hex_codec")
-logger = logging.getLogger(__name__)
-
-# TODO: refactor this file structure
 
 
-class ROAParserException(Exception):
-    pass
-
-
-class ROADataImporter:
-    def __init__(self, text: str, database_handler: DatabaseHandler):
-        self.roa_objs: List[ROA] = []
-        try:
-            roa_dicts = ujson.loads(text)['roas']
-        except ValueError as error:
-            msg = f'Unable to parse ROA input: invalid JSON: {error}'
-            logger.error(msg)
-            raise ROAParserException(msg)
-        except KeyError:
-            msg = f'Unable to parse ROA input: root key "roas" not found'
-            logger.error(msg)
-            raise ROAParserException(msg)
-
-        for roa_dict in roa_dicts:
-            try:
-                roa_obj = ROA(
-                    roa_dict['prefix'],
-                    roa_dict['asn'],
-                    roa_dict['maxLength'],
-                    roa_dict['ta'],
-                )
-            except KeyError as ke:
-                msg = f'Unable to parse ROA record: missing key {ke} -- full record: {roa_dict}'
-                logger.error(msg)
-                raise ROAParserException(msg)
-
-            roa_obj.save(database_handler)
-            self.roa_objs.append(roa_obj)
-
-
-class ROA:
-    def __init__(self, prefix: str, asn: str, max_length: str, trust_anchor: str):
-        try:
-            self.prefix = IP(prefix)
-            self.prefix_str = prefix
-            _, self.asn = parse_as_number(asn)
-            self.max_length = int(max_length)
-            self.trust_anchor = trust_anchor
-        except ValueError as ve:
-            msg = f'Invalid value in ROA: {ve}'
-            logger.error(msg)
-            raise ROAParserException(msg)
-
-        if self.max_length < self.prefix.prefixlen():
-            msg = f'Invalid ROA: prefix size {self.prefix.prefixlen()} is smaller than max length {max_length} in ' \
-                  f'ROA for {self.prefix} / AS{self.asn}'
-            logger.error(msg)
-            raise ROAParserException(msg)
-
-    def save(self, database_handler: DatabaseHandler):
-        database_handler.insert_roa_object(
-            ip_version=self.prefix.version(),
-            prefix_str=self.prefix_str,
-            asn=self.asn,
-            max_length=self.max_length,
-            trust_anchor=self.trust_anchor,
-        )
-        self._rpsl_object = RPSLObjectFromROA(
-            prefix=self.prefix,
-            prefix_str=self.prefix_str,
-            asn=self.asn,
-            max_length=self.max_length,
-            trust_anchor=self.trust_anchor,
-        )
-        database_handler.upsert_rpsl_object(self._rpsl_object, rpsl_safe_insert_only=True)
-
-
-class RPSLObjectFromROA(RPSLObject):
-    def __init__(self, prefix: IP, prefix_str: str, asn: int, max_length: int, trust_anchor: str):
-        self.prefix = prefix
-        self.prefix_str = prefix_str
-        self.asn = asn
-        self.max_length = max_length
-        self.trust_anchor = trust_anchor
-
-        self.rpsl_object_class = RPSL_ROUTE_OBJECT_CLASS_FOR_IP_VERSION[self.prefix.version()]
-        self.ip_first = self.prefix.net()
-        self.ip_last = self.prefix.broadcast()
-        self.prefix_length = self.prefix.prefixlen()
-        self.asn_first = asn
-        self.asn_last = asn
-        self.parsed_data = {
-            self.rpsl_object_class: self.prefix_str,
-            'origin': 'AS' + str(self.asn),
-            'source': RPKI_IRR_PSEUDO_SOURCE,
-        }
-
-    def source(self):
-        return RPKI_IRR_PSEUDO_SOURCE
-
-    def pk(self):
-        return f'{self.prefix_str}AS{self.asn}/ML{self.max_length}'
-
-    def render_rpsl_text(self):
-        # TODO: we could just have a max-length attribute?
-        object_class_display = f'{self.rpsl_object_class}:'.ljust(RPSL_ATTRIBUTE_TEXT_WIDTH)
-        rpsl_object_text = textwrap.dedent(f"""
-            {object_class_display}{self.prefix_str}
-            descr:          RPKI ROA for {self.prefix_str} / AS{self.asn}
-            remarks:        This route object represents routing data retrieved from the RPKI
-            remarks:        The original data can be found here: https://rpki.gin.ntt.net/r/AS{self.asn}/{self.prefix_str}
-            remarks:        This route object is the result of an automated RPKI-to-IRR conversion process
-            remarks:        performed by IRRd.
-            remarks:        maxLength {self.max_length}
-            origin:         AS{self.asn}
-            source:         {RPKI_IRR_PSEUDO_SOURCE}  # Trust Anchor: {self.trust_anchor}
-            """).strip() + '\n'
-        return rpsl_object_text
-
-
-class BulkRouteRoaValidator:
+class BulkRouteROAValidator:
     """
     The bulk route validator is optimised to validate large amounts
     of routes, e.g. all RPSL route(6)s in the DB.
@@ -148,8 +25,8 @@ class BulkRouteRoaValidator:
     into that tree. For each route to be validated, the tree is searched
     to find any matching ROAs.
 
-    ROAs are inserted under a key that is a binary representation of their
-    prefix, including only network bits. For 192.0.2.0/24, the binary
+    ROAs are inserted under a key that is a binary string representation of
+    their prefix, including only network bits. For 192.0.2.0/24, the binary
     representation is 11000000000000000000001000000000, reducing this to
     network bits mean the key is 110000000000000000000010.
     As multiple ROAs may exist for the same prefix, the data under the
@@ -157,18 +34,21 @@ class BulkRouteRoaValidator:
     tuple representing a ROA.
 
     When searching for ROAs for a route, the entire network address
-    is converted to binary. E.g. for 192.0.2.139, this would be
-    11000000000000000000001010001011. The trie is then searched for
-    any items that are a prefix of 11000000000000000000001010001011,
-    which includes 110000000000000000000010, the key for 192.0.2(.0).
+    is converted to binary, and also trimmed to network bits.
+    E.g. for 192.0.2.0/25, this would be 1100000000000000000000100.
+    The trie is then searched for any items that are a prefix (inclusive) of
+    1100000000000000000000100, which includes 110000000000000000000010,
+    the key for 192.0.2(.0). A ROA for 192.0.2.0/26 would have key
+    11000000000000000000001000, and therefore will (and should)
+    not be included in the validation process.
     """
     def __init__(self, dh: DatabaseHandler, roas: Optional[List[ROA]] = None):
         """
-        Create a validator object. Can use either a list of ROA bojects,
+        Create a validator object. Can use either a list of ROA objects,
         or if not give, generates this from the database.
         Due to the overhead in preloading all ROAs, this is only effective
         when many routes have to be validated, otherwise it's more
-        efficient to use SingleRouteRoaValidator. The break even
+        efficient to use SingleRouteROAValidator. The break even
         point is in the order of magnitude of checking 10.000 routes.
         """
         self.database_handler = dh
@@ -204,16 +84,15 @@ class BulkRouteRoaValidator:
         pks_changed: Dict[RPKIStatus, Set[str]] = defaultdict(set)
 
         for result in routes:
-            # RPKI_IRR_PSEUDO_SOURCE objects are ROAs,
-            # and don't need validation.
+            # RPKI_IRR_PSEUDO_SOURCE objects are ROAs, and don't need validation.
             if result['source'] == RPKI_IRR_PSEUDO_SOURCE:
                 continue
 
             rpsl_pk = result['rpsl_pk']
             current_status = result['rpki_status']
-            status = self.validate_route(result['ip_first'], result['prefix_length'], result['asn_first'])
-            if status != current_status:
-                pks_changed[status].add(rpsl_pk)
+            new_status = self.validate_route(result['ip_first'], result['prefix_length'], result['asn_first'])
+            if new_status != current_status:
+                pks_changed[new_status].add(rpsl_pk)
 
         return pks_changed[RPKIStatus.valid], pks_changed[RPKIStatus.invalid], pks_changed[RPKIStatus.unknown]
 
@@ -275,11 +154,11 @@ class BulkRouteRoaValidator:
         """
         address_family = socket.AF_INET6 if ':' in ip else socket.AF_INET
         ip_bin = socket.inet_pton(address_family, ip)
-        ip_bin_str = ''.join([self._byte_bin[x] for x in ip_bin]) + '0'
+        ip_bin_str = ''.join([self._byte_bin[b] for b in ip_bin]) + '0'
         return ip_bin_str
 
 
-class SingleRouteRoaValidator:
+class SingleRouteROAValidator:
     def __init__(self, database_handler: DatabaseHandler):
         self.database_handler = database_handler
