+++ irrd/rpsl/parser.py
@@ -58,7 +58,7 @@ class RPSLObject(metaclass=RPSLObjectMeta):
     ip_last: IP = None
     asn_first: Optional[int] = None
     asn_last: Optional[int] = None
-    prefix: IP = None
+    prefix: IP = None  # Note: not saved to DB, used for performance
     prefix_length: Optional[int] = None
     rpki_status: RPKIStatus = RPKIStatus.unknown
     default_source: Optional[str] = None  # noqa: E704 (flake8 bug)
